+++ src/components/Logo/index.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Link } from 'gatsby'
+import logo from '../../images/icon.png'
+
+const Logo = () => (
+  <Link to="/" style={{
+    marginRight: '20px'
+  }}>
+    <img src={logo} alt="" height="40" />
+  </Link>
+)
+
+export default Logo;
