+++ src/components/SiteHeader/index.js
@@ -3,24 +3,15 @@ import React, { Component } from 'react'
 import {
   Layout, Menu, Row, Button, Icon
 } from 'antd'
-import Container from '../Container'
 import { StaticQuery, graphql, navigate, Link } from 'gatsby'
 import SiteSearch from '../SiteSearch'
-import logo from '../../images/icon.png'
 
 const ButtonGroup = Button.Group
 
 const {
-  Header,
 } = Layout
 
-const Logo = () => (
-    <Link to="/" style={{
-      marginRight: '20px'
-    }}>
-      <img src={logo} alt="" height="40" />
-    </Link>
-)
+
 
 const menuItems = [
   {
@@ -37,57 +28,6 @@ const menuItems = [
     path: '/blog'
   }
 ];
-
-class SiteMenu extends Component {
-
-  getSelectedKeys = () => {
-    const { location } = this.props;
-
-    if ( ! location || ! location.pathname ) {
-      return null;
-    }
-
-    if ( location.pathname.includes( '/docs') ) {
-      return ['/docs']
-    }
-
-    if ( location.pathname.includes( '/community') ) {
-      return ['/community']
-    }
-
-    if ( location.pathname.includes( '/blog') ) {
-      return ['/blog']
-    }
-
-    return [];
-
-  }
-
-  render() {
-    return (
-      <div style={{
-        maxWidth: '600px',
-      }}>
-        <Menu
-          theme="dark"
-          mode="horizontal"
-          defaultSelectedKeys={['/docs']}
-          selectedKeys={this.getSelectedKeys()}
-          style={{ lineHeight: '64px', borderBottom: '0px' }}
-        >
-          {menuItems.map(item => (
-            <Menu.Item
-              key={ item.key ? item.key : item.path}
-              onClick={() => navigate(item.path)}
-            >
-              {item.name}
-            </Menu.Item>
-          ))}
-        </Menu>
-      </div>
-    )
-  }
-}
 
 const UtilNav = () => {
   return (
@@ -134,24 +74,11 @@ const UtilNav = () => {
 }
 
 const SiteHeader = ({ siteTitle, location }) => (
-  <Header style={{ padding: 0, height: '66px', position: 'fixed', zIndex: 101, width: '100%' }}>
-    <Container>
-      <Row type="flex" justify="space-between">
-        <div className="logo-menu">
-          <Row type="flex" justify="start">
-            <Logo siteTitle={siteTitle}/>
-            <SiteMenu location={location} />
-          </Row>
-        </div>
-        <div className="search-util">
-          <Row type="flex" justify="end">
-            <SiteSearch />
-            <UtilNav/>
-          </Row>
-        </div>
-      </Row>
-    </Container>
-  </Header>
+
+    <Row type="flex" justify="end">
+      <SiteSearch />
+      <UtilNav/>
+    </Row>
 )
 
 SiteHeader.propTypes = {
