+++ src/components/SiteSearch/index.js
@@ -31,7 +31,7 @@ class Complete extends Component {
   render() {
     const { results } = this.state;
     return (
-      <div className="certain-category-search-wrapper" style={{ width: '100%', padding: '1px 20px' }}>
+      <div className="certain-category-search-wrapper" style={{ width: '400px', padding: '1px 20px' }}>
         <AutoComplete
           className="certain-category-search"
           dropdownClassName="certain-category-search-dropdown"
